Move property UDFs to datafusion-extensions for WASM support
Move PropertyGet, PropertiesToArray, and PropertiesLength to
micromegas-datafusion-extensions so the browser query engine
gets property support. PropertiesToDict and PropertiesToJsonb
stay server-side in analytics.
